AsientosSalas terminado

+++ src/main/java/com/entradaservice/repository/SalaRepository.java
@@ -4,6 +4,9 @@ import com.entradaservice.models.Sala;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface SalaRepository extends JpaRepository<Sala, Long> {
+public interface SalaRepository extends JpaRepository<Sala, Integer> {
+
 }
